Implement Unary RPC
Urinary RPC is where a client sends a single request to the server and get a single response from the server synchronously. We have a client that calls our RPC 'createTodo' pass in the do that want to create the server response back with the id incremented and the text. eg {"id":3,"text":"Do Laundry"} becomes {"id":4,"text":"Do Laundry"}

In the server, we get this when it receive a request from client
ServerUnaryCall {
  _events: [Object: null prototype] { error: [Function (anonymous)] },
  _eventsCount: 1,
  _maxListeners: undefined,
  call: Call { metadataSent: true },
  cancelled: false,
  metadata: Metadata { _internal_repr: { 'user-agent': [Array] }, flags: 0 },
  request: { id: -1, text: 'Do Laundry' },
  [Symbol(kCapture)]: false
}
NOTE: the metadata is similar concept as HTTP headers for restAPI

FYI: the package '@grpc/proto-loader' automatically converts the protoFile into js script files that generates our schema, getters and setters. So this is good stuff. We don't have to do that ourself.

You will get warnings for not implementing other RPC like readTodos and readTodoStream but that ok. we will implement those down the line

The client and server have to be on thesame port. Our serve will reconnect to any ip address from client as far as it connects on the port