Implement Server RPC (Non-Streamed)
This is when a client sends a request to the sever, but the server responds back with a lot of data.

FYI: the method we used here is not a best practise because it is not streamed.

The client receives something like this from the server
{"items":[{"id":1,"text":"gym"},{"id":2,"text":"gyming"},{"id":3,"text":"run"},{"id":4,"text":"running"}]}
You access this by response.items then you can loop through the or do whatever