Implement Client rPC (uploading image)
Client streaming is when the client sends a lot of request with the server sending one response back like uploading files or sending an order list record to save to the server

In this commit, we go a file and streamed the file chunk by chunk to the server. The we save the file to a folder. OfCourse no, unique name for the file. But we can do that for ideal situation

The key here is to client stream an image using call.write and the server uses call.on eventListener to get all the stream. So server has event listener and client does not